requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.18.5
pandas==1.3.5
scikit-learn==0.22

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs==2022.5.0
google-cloud-storage==2.4.0
mlflow
s3fs

# utilities
six==1.15.0
joblib
memoized-property
termcolor
